+++ Deribit/Core/Messages/Trading/EditMessage.cs
@@ -13,13 +13,13 @@ namespace Deribit.Core.Messages.Trading
         }
 
         public string order_id { get; set; }
-        public float amount { get; set; }
-        public float? price { get; set; }
+        public decimal amount { get; set; }
+        public decimal? price { get; set; }
         public bool? post_only { get; set; }
         public bool? reduce_only { get; set; }
         public bool? reject_post_only { get; set; }
         public string? advanced { get; set; }
-        public float? stop_price { get; set; }
+        public decimal? stop_price { get; set; }
         public bool? mmp { get; set; }
         
         public void CheckValidity()
