+++ Deribit/Core/Notifications/NotificationObjects/SubscriptionData.cs
@@ -3,7 +3,7 @@
     public class SubscriptionData
     {
         public long timestamp { get; set; }
-        public float price { get; set; }
+        public decimal price { get; set; }
         public string index_name { get; set; }
     }
 }
